+++ src/Handler/Combination/CombinationExportHandler.php
@@ -10,7 +10,10 @@ use FactorioItemBrowser\Api\Client\Request\Combination\CombinationExportRequest;
 use FactorioItemBrowser\Api\Client\Request\RequestInterface as ClientRequestInterface;
 use FactorioItemBrowser\Api\Client\Response\Combination\CombinationStatusResponse;
 use FactorioItemBrowser\Api\Client\Response\ResponseInterface as ClientResponseInterface;
+use FactorioItemBrowser\Api\Server\Entity\AuthorizationToken;
+use FactorioItemBrowser\Api\Server\Exception\ActionNotAllowedException;
 use FactorioItemBrowser\Api\Server\Handler\AbstractRequestHandler;
+use FactorioItemBrowser\Api\Server\Service\AgentService;
 use FactorioItemBrowser\Api\Server\Service\ExportQueueService;
 use FactorioItemBrowser\ExportQueue\Client\Constant\JobStatus;
 use FactorioItemBrowser\ExportQueue\Client\Exception\ClientException;
@@ -24,6 +27,12 @@ use FactorioItemBrowser\ExportQueue\Client\Exception\ClientException;
 class CombinationExportHandler extends AbstractRequestHandler
 {
     /**
+     * The agent service.
+     * @var AgentService
+     */
+    protected $agentService;
+
+    /**
      * The export queue service.
      * @var ExportQueueService
      */
@@ -31,10 +40,12 @@ class CombinationExportHandler extends AbstractRequestHandler
 
     /**
      * Initializes the handler.
+     * @param AgentService $agentService
      * @param ExportQueueService $exportQueueService
      */
-    public function __construct(ExportQueueService $exportQueueService)
+    public function __construct(AgentService $agentService, ExportQueueService $exportQueueService)
     {
+        $this->agentService = $agentService;
         $this->exportQueueService = $exportQueueService;
     }
 
@@ -51,12 +62,16 @@ class CombinationExportHandler extends AbstractRequestHandler
      * Creates the response data from the validated request data.
      * @param ClientRequestInterface $clientRequest
      * @return ClientResponseInterface
+     * @throws ActionNotAllowedException
      * @throws ClientException
      * @throws MapperException
      */
     protected function handleRequest($clientRequest): ClientResponseInterface
     {
         $authorizationToken = $this->getAuthorizationToken();
+        if (!$this->isAgentAllowed($authorizationToken)) {
+            throw new ActionNotAllowedException();
+        }
 
         [$latestJobResponse, $latestSuccessfulJobResponse] = $this->exportQueueService->executeListRequests([
             $this->exportQueueService->createListRequest($authorizationToken),
@@ -77,6 +92,17 @@ class CombinationExportHandler extends AbstractRequestHandler
                  ->setLatestExportJob($latestExportJob)
                  ->setLatestSuccessfulExportJob($latestSuccessfulExportJob);
         return $response;
+    }
+
+    /**
+     * Returns whether the current agent is actually allowed to create exports.
+     * @param AuthorizationToken $authorizationToken
+     * @return bool
+     */
+    protected function isAgentAllowed(AuthorizationToken $authorizationToken): bool
+    {
+        $agent = $this->agentService->getByName($authorizationToken->getAgentName());
+        return $agent !== null && !$agent->getIsDemo();
     }
 
     /**